pyproject.toml
[project]
name = "open-blood-analysis"
version = "0.0.1"
description = "Open Blood Analysis"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "filelock>=3.20.3",
    "pdf2image>=1.17.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.0.0",
    "rapidfuzz>=3.14.3",
    "typer>=0.21.1",
    "simpleeval>=0.9.13",
    "google-genai>=1.0.0",
    "pytest>=9.0.2",
]

[project.scripts]
blood-analysis = "app.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
